requirements.txt
altgraph==0.17
astroid==2.4.2
certifi==2020.12.5
colorama==0.4.4
cycler==0.10.0
future==0.18.2
graphviz==0.16
guidata==1.7.9
guiqwt==3.0.6
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.4.2
mccabe==0.6.1
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1618854156666/work
mkl-service==2.3.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1618497418457/work
olefile==0.46
pefile==2021.5.13
Pillow @ file:///C:/ci/pillow_1617386341487/work
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pylint==2.6.0
pyparsing==2.4.7
pyqt-tools==1.0.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
PythonQwt==0.9.0
pywin32-ctypes==0.2.0
qimage2ndarray==1.8.3
QtPy==1.9.0
scipy==1.6.3
six==1.15.0
toml==0.10.2
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
torchviz==0.0.2
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
wincertstore==0.2
wrapt==1.12.1
